BudgetDto, LaborBudgetDto and MaterialBudgetDto keywords edited.

+++ src/main/java/tug/it/openprojectapi/model/LaborBudgetDto.java
@@ -13,5 +13,5 @@ public class LaborBudgetDto {
     private Integer id;
     private Integer user_id;
     private Double hours;
-    private Integer budget;
+    private Double budget;
 }
